[feature] implement new playbook

+++ src/playbook/model/playbook.rs
@@ -4,14 +4,14 @@ use crate::schema::playbook;
 pub struct Playbook {
     pub id: i64,
     pub title: String,
-    pub create_time: i32,
-    pub update_time: i32,
+    pub create_time: i64,
+    pub update_time: i64,
 }
 
 #[derive(Insertable)]
 #[table_name="playbook"]
 pub struct NewPlaybook {
     pub title: String,
-    pub create_time: i32,
-    pub update_time: i32,
+    pub create_time: i64,
+    pub update_time: i64,
 }
