+++ src/schema.rs
@@ -2,7 +2,7 @@ table! {
     playbook (id) {
         id -> Nullable<Integer>,
         title -> Nullable<Text>,
-        create_time -> Nullable<Integer>,
-        update_time -> Nullable<Integer>,
+        create_time -> Nullable<BigInt>,
+        update_time -> Nullable<BigInt>,
     }
 }
